% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coloc_show}
\alias{coloc_show}
\title{Show the colocalization output}
\usage{
coloc_show(corr)
}
\arguments{
\item{corr}{A list, such as that returned by coloc_test with num == TRUE}
}
\description{
Show the colocalization output
}
\examples{
# load libraries
library(imager)

# load images
fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
img <- load.image(fl)

# choose parameters
px <- roi_select(img, threshold = 90)

# call coloc_test
corr <- coloc_test(img, px, num = TRUE)

# call coloc_show
coloc_show(corr)

}
