% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coloc_test}
\alias{coloc_test}
\title{Test Co-localization}
\usage{
coloc_test(img, px, type = "pearsons", num = FALSE, ind = c(1, 2))
}
\arguments{
\item{img}{An object of class \code{\link[imager]{cimg}}}

\item{px}{An object of class \code{\link[imager]{pixset}} or with labels
\code{\link[imager]{cimg}}}

\item{type}{A \code{character}; "pearsons", "manders" or "all"}

\item{num}{A \code{logical}; return the \code{numeric} values of the images
or not}

\item{ind}{A \code{numeric} of length two for channel indecies}
}
\value{
A \code{list} of one or two correlations measures and two
\code{numeric}s when num is TRUE
}
\description{
Perform co-localization test statistics.
}
\examples{
# load libraries
library(imager)

# load images
fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
img <- load.image(fl)

# choose parameters
px <- roi_select(img, threshold = 90)

# call coloc_test
coloc_test(img, px)

}
