% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocr.R
\docType{package}
\name{colocr}
\alias{colocr}
\alias{colocr-package}
\title{\code{colocr}: Conduct Co-localization Analysis of Microscopy Images.}
\description{
Automate the co-localization analysis of fluorescence microscopy images.
Selecting regions of interest, extract pixel intensities from the image
channels and calculate different co-localization statistics.
}
\section{\code{colocr} functions}{

\code{\link{roi_select}}
\code{\link{roi_show}}
\code{\link{roi_check}}
\code{\link{roi_test}}
}

