% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.labels_add}
\alias{.labels_add}
\title{Label regions of interest}
\usage{
.labels_add(px, tolerance, n)
}
\arguments{
\item{px}{An object of class \code{\link[imager]{pixset}}}

\item{tolerance}{A \code{numeric} to be passed to \code{\link[imager]{label}}}

\item{n}{A \code{numeric}, the number of desired regions of interest}
}
\value{
An object of class \code{\link[imager]{cimg}}. The labels are coded
the values in the object starting from 1. The rest of the image is labeled 0.
}
\description{
Add labels to regions of interest in an image
}
