% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.pearson}
\alias{.pearson}
\title{Calculate Pearson's Correlation Coefficient}
\usage{
.pearson(r, g)
}
\arguments{
\item{r}{A \code{numeric} vector}

\item{g}{A \code{numeric} vector}
}
\value{
A \code{numeric} of length one.
}
\description{
Calculates the Pearson's correlation coefficient between two numeric vectors
}
\examples{
set.seed(123)
r <- rnorm(10)

set.seed(1234)
g <- rnorm(10)

.pearson(r, g)

}
