% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{image_load}
\alias{image_load}
\title{Load images from files}
\usage{
image_load(image_file)
}
\arguments{
\item{image_file}{A \code{character} vector of one or more paths to image
files}
}
\value{
A \code{cimg} object or a \code{list} of \code{cimg} objects when
multiple files are passed to \code{image_file}.
}
\description{
A wrap around \code{\link[magick]{image_read}} and
\code{\link[imager]{magick2cimg}} to load one or more images from files.
}
\examples{
# load image
fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
img <- image_load(fl)

}
