% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{intensity_get}
\alias{intensity_get}
\title{Get pixel intensities}
\usage{
intensity_get(img, px, ind = c(1, 2))
}
\arguments{
\item{img}{An object of class \code{\link[imager]{cimg}}}

\item{px}{An object of class \code{\link[imager]{pixset}} or with labels
\code{\link[imager]{cimg}}}

\item{ind}{A \code{numeric} of length two for channel indecies}
}
\value{
A \code{list} of three items. The first two items are the values of
the pixel intensities of the channels indicated by \code{ind}. The third is
the labels of the individual regions of interest when a labeled \code{px}
is provided. Otherwise, the value of labels is set to 1.
}
\description{
Get the pixel intensities of certain image channels
}
\examples{
# load libraries
library(imager)

# load images
fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
img <- load.image(fl)

# choose parameters
px <- roi_select(img, threshold = 90)

# call coloc_test
pix_int <- intensity_get(img, px)

}
