% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{labels_add}
\alias{labels_add}
\title{Make labels}
\usage{
labels_add(px, tolerance = 0.1, n = 1)
}
\arguments{
\item{px}{An object of class \code{\link[imager]{pixset}}}

\item{tolerance}{A \code{numeric} to be passed to \code{\link[imager]{label}}}

\item{n}{A \code{numeric} of the number of regions of interest}
}
\value{
An object of class \code{\link[imager]{cimg}}
}
\description{
Make labels
}
