% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_choose.R
\name{parameter_choose}
\alias{parameter_choose}
\title{Choose Parameter for selecting ROI}
\usage{
parameter_choose(img, threshold, shrink = 5, grow = 5, fill = 5,
  clean = 5)
}
\arguments{
\item{img}{An object of class \code{\link[imager]{cimg}}}

\item{threshold}{A numeric to be passed to threshold}

\item{shrink}{A numeric to be passed to shrink}

\item{grow}{A numeric to be passed to grow}

\item{fill}{A numeric to be passed to fill}

\item{clean}{A numeric to be passed to clean}
}
\value{
A pixset
}
\description{
Choose the different parameter to select the regions of interest (ROI)
}
\examples{
# load libraries
library(imager)
library(colocr)

# load images
fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
img <- load.image(fl)
img.g <- grayscale(img)

# choose parameters
px <- parameter_choose(img.g, threshold = 90)

# highlight chosen region of interest
plot(img)
highlight(px)

}
