% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameter_show}
\alias{parameter_show}
\title{Show the selected ROIs}
\usage{
parameter_show(img, px, labels, ind = c(1, 2))
}
\arguments{
\item{img}{An object of class \code{\link[imager]{cimg}}}

\item{px}{An object of class \code{\link[imager]{pixset}}}

\item{labels}{An object of class \code{\link[imager]{cimg}}}

\item{ind}{A \code{numeric} of length two for channel indecies}
}
\description{
Show/highlight the selected ROIs for a set of parameters
}
\examples{
# load libraries
library(imager)

# load images
fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
img <- load.image(fl)

# choose parameters
px <- parameter_choose(img, threshold = 90)

# call coloc_test
parameter_show(img, px)

}
