% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{roi_show}
\alias{roi_show}
\title{Show the selected regions of interest}
\usage{
roi_show(img, px, labels, ind = c(1, 2))
}
\arguments{
\item{img}{An object of class \code{\link[imager]{cimg}}}

\item{px}{An object of class \code{\link[imager]{pixset}}}

\item{labels}{An object of class \code{\link[imager]{cimg}}}

\item{ind}{A \code{numeric} object of length two. For the channel indexes.}
}
\description{
Show/highlight the selected regions of interest on different image channels
}
\details{
Calling this function on an image and a \code{px}, \code{\link[imager]{pixset}}
of the same dimensions returns four different plots. The original image, a
a low resolution representation of the \code{\link[imager]{pixset}} and the
two channels indicated through \code{ind} highlighted. Additionally, when
labels are provided through \code{labels} the regions are individually
labeled.
}
\examples{
# load libraries
library(imager)

# load images
fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
img <- load.image(fl)

# choose parameters
px <- roi_select(img, threshold = 90)

# call coloc_test
roi_show(img, px)

}
