% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{roi_test}
\alias{roi_test}
\title{Test Co-localization}
\usage{
roi_test(img, ind = c(1, 2), type = "pcc")
}
\arguments{
\item{img}{A \code{\link[imager]{cimg}} obeject such as the one returned
from \code{\link{roi_select}}}

\item{ind}{A \code{numeric} object of length two. For the channel indexes.}

\item{type}{A \code{character} vector of the desired co-localization
statistics. Default is 'pcc', other inputs are 'moc' or 'both'.}
}
\value{
A \code{list}.
}
\description{
Perform co-localization test statistics.
}
\details{
The co-localization stats requested in \code{type} is returned as
list items for each. When different labels are provided, the stats are
calculated for each label individually.
}
\examples{
# load required libraries
library(imager)
library(magrittr)

# load images
fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
img <- load.image(fl)

# choose roi and test colocalization
roi_select(img, threshold = 90) \%>\%
  roi_test()

}
