% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{coloc_show}
\alias{coloc_show}
\title{Show the colocalization output}
\usage{
coloc_show(corr)
}
\arguments{
\item{corr}{A list, such as that returned by coloc_test with num == TRUE}
}
\description{
Show the colocalization output
}
\examples{
# load libraries
library(imager)

# load images
fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
img <- load.image(fl)
img.g <- grayscale(img)

# choose parameters
px <- parameter_choose(img.g, threshold = 90)

# load images from two channels
img1 <- load.image(system.file('extdata/', 'Image0001_C002.jpg', package = 'colocr'))
img2 <- load.image(system.file('extdata/', 'Image0001_C003.jpg', package = 'colocr'))

# call coloc_test
corr <- coloc_test(img1, img2, px, num = TRUE)

# call coloc_show
coloc_show(corr)

}
