% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc.R
\name{coloc_test}
\alias{coloc_test}
\title{Test colocalization}
\usage{
coloc_test(img1, img2, px, type = "pearsons", num = FALSE)
}
\arguments{
\item{img1}{An object of class cimg}

\item{img2}{An object of class cimg}

\item{px}{An object of class pixset}

\item{type}{A character; pearson's, spearman or all}

\item{num}{A logical; return the numeric values of the images or not.}
}
\value{
A list of one or two correlations measures and two numerics when num
is TRUE
}
\description{
Test colocalization
}
\examples{
# load libraries
library(imager)

# load images
fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
img <- load.image(fl)
img.g <- grayscale(img)

# choose parameters
px <- parameter_choose(img.g, threshold = 90)

# load images from two channels
img1 <- load.image(system.file('extdata/', 'Image0001_C002.jpg', package = 'colocr'))
img2 <- load.image(system.file('extdata/', 'Image0001_C003.jpg', package = 'colocr'))

# call coloc_test
coloc_test(img1, img2, px)

}
