% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coloc_test}
\alias{coloc_test}
\title{Test Co-localization}
\usage{
coloc_test(pix_int, type = "pearsons")
}
\arguments{
\item{pix_int}{A list, such as that returned by \code{\link{intensity_get}}}

\item{type}{A \code{character}; "pearsons", "manders" or "all"}
}
\value{
A \code{list}.
}
\description{
Perform co-localization test statistics.
}
\details{
The co-localization stats requested in \code{type} is returned as
list items for each. When different labels are provided, the stats are
calculated for each label individually.
}
\examples{
# load libraries
library(imager)

# load images
fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
img <- load.image(fl)

# choose roi
px <- roi_select(img, threshold = 90)

# get intensities
pix_int <- intensity_get(img, px)

# call coloc_test
coloc_test(pix_int)

}
