% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{intensity_show}
\alias{intensity_show}
\title{Show pixel intensities}
\usage{
intensity_show(pix_int)
}
\arguments{
\item{pix_int}{A list, such as that returned by \code{\link{intensity_get}}}
}
\description{
Show the pixel intensities of certain image channels
}
\details{
Calling this function returns two plots. The first is a scatter
plot of the pixel intensities from two channels. The second is the density
distribution of the intensities from the two channels.
}
\examples{
# load libraries
library(imager)

# load images
fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
img <- load.image(fl)

# choose parameters
px <- roi_select(img, threshold = 90)

# call coloc_test
pix_int <- intensity_get(img, px)

# call intensity_show
intensity_show(pix_int)

}
