% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameter_choose}
\alias{parameter_choose}
\title{Choose Parameter for selecting ROIs}
\usage{
parameter_choose(img, threshold, shrink = 5, grow = 5, fill = 5,
  clean = 5)
}
\arguments{
\item{img}{An object of class \code{\link[imager]{cimg}}}

\item{threshold}{A \code{numeric} to be passed to
\code{\link[imager]{threshold}}}

\item{shrink}{A \code{numeric} to be passed to \code{\link[imager]{shrink}}}

\item{grow}{A \code{numeric} to be passed to \code{\link[imager]{grow}}}

\item{fill}{A \code{numeric} to be passed to \code{\link[imager]{fill}}}

\item{clean}{A \code{numeric} to be passed to \code{\link[imager]{clean}}}
}
\value{
A \code{\link[imager]{pixset}}
}
\description{
Choose the different parameters to select the regions of interest (ROI)
}
\examples{
# load libraries
library(imager)
library(colocr)

# load images
fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
img <- load.image(fl)

# choose parameters
px <- parameter_choose(img, threshold = 90)

# highlight chosen region of interest
plot(img)
highlight(px)

}
