% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameter_show}
\alias{parameter_show}
\title{Show the selected ROIs}
\usage{
parameter_show(img, img1, img2, px, labels)
}
\arguments{
\item{img}{An object of class \code{\link[imager]{cimg}}}

\item{img1}{An object of class \code{\link[imager]{cimg}}}

\item{img2}{An object of class \code{\link[imager]{cimg}}}

\item{px}{An object of class \code{\link[imager]{pixset}}}

\item{labels}{A \code{data.frame} of three columns; x, y and value.}
}
\description{
Show/highlight the selected ROIs for a set of parameters
}
\examples{
# load libraries
library(imager)

# load images
fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
img <- load.image(fl)
img.g <- grayscale(img)

# choose parameters
px <- parameter_choose(img.g, threshold = 90)

# load images from two channels
img1 <- load.image(system.file('extdata/', 'Image0001_C002.jpg', package = 'colocr'))
img2 <- load.image(system.file('extdata/', 'Image0001_C003.jpg', package = 'colocr'))

# call coloc_test
parameter_show(img, img1, img2, px)

}
