% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{roi_select}
\alias{roi_select}
\title{Select regions of interest}
\usage{
roi_select(img, threshold, shrink = 5, grow = 5, fill = 5, clean = 5,
  tolerance = 0.1, n)
}
\arguments{
\item{img}{An object of class \code{\link[imager]{cimg}}}

\item{threshold}{A \code{numeric} to be passed to
\code{\link[imager]{threshold}}}

\item{shrink}{A \code{numeric} to be passed to \code{\link[imager]{shrink}}}

\item{grow}{A \code{numeric} to be passed to \code{\link[imager]{grow}}}

\item{fill}{A \code{numeric} to be passed to \code{\link[imager]{fill}}}

\item{clean}{A \code{numeric} to be passed to \code{\link[imager]{clean}}}

\item{tolerance}{A \code{numeric} to be passed to \code{\link[imager]{label}}}

\item{n}{A \code{numeric} of the number of regions of interest}
}
\value{
A \code{\link[imager]{pixset}} or a labeled image \code{\link[imager]{cimg}}
}
\description{
Select regions of interest in an image using different morphological operations
}
\details{
The function applies several \code{\link{imager}} morphological
manipulations to select the regions of interest. Thes include
\code{\link[imager]{threshold}} which sets all values below certain cut to
0; \code{\link[imager]{shrink}}/\code{\link[imager]{grow}} for pixel set
dilation and erosion; \code{\link[imager]{fill}}\\code{\link[imager]{clean}}
for removing isolated regions and holes. When \code{n} is provided, the
individual regions (connected components) are selected where \code{tolerance}
is used to determin if two pixels belong to the same region.
}
\examples{
# load libraries
library(imager)
library(colocr)

# load images
fl <- system.file('extdata', 'Image0001_.jpg', package = 'colocr')
img <- load.image(fl)

# choose parameters
px <- roi_select(img, threshold = 90)

# highlight chosen region of interest
plot(img)
highlight(px)

}
